depolyment/requirements.txt
numpy>=1.21
Pillow>=9.0
tflite-runtime>=2.12

